requirements.txt
Flask==3.0.3
mysql-connector-python==9.0.0
PySide6==6.7.2
requests==2.32.3
python-dotenv==1.0.1
bcrypt==4.2.0

